requirements.txt
langchain>=1.0.8
langchain-community>=0.4.1
langchain-core>=1.1.0
langchain-text-splitters>=1.0.0
langchain-classic>=1.0.0
chromadb>=0.4.22
ollama>=0.1.7
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.1
python-dotenv>=1.0.0
pytest>=7.4.4
pytest-cov>=4.1.0
unstructured>=0.12.0
pygments>=2.17.2